+++ click_extra/__init__.py
@@ -43,7 +43,7 @@ from .commands import (
     ExtraGroup,
     LazyGroup,
 )
-from .config import ConfigFormat, ConfigOption, NoConfigOption
+from .config import ConfigFormat, ConfigOption, NO_CONFIG, NoConfigOption
 from .decorators import (  # type: ignore[no-redef]
     argument,
     color_option,
@@ -161,6 +161,7 @@ __all__ = [  # noqa: F405
     "make_pass_decorator",
     "MissingParameter",
     "new_extra_logger",
+    "NO_CONFIG",
     "no_config_option",
     "NoConfigOption",
     "NoSuchOption",